// workspaces/ballerina/ballerina-core/src/interfaces/ballerina.ts
export interface ComponentInfo {
    name: string;
    filePath: string;
    startLine: number;
    startColumn: number;
    endLine: number;
    endColumn: number;
    resources?: ComponentInfo[];
}